data layer manager as revealed function in separate module
